PythonBackend/Vegaani_App/requirements_personal.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
distlib==0.3.1
Django==3.1.2
django-bootstrap-form==3.4
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-filter==2.4.0
djangorestframework==3.12.1
filelock==3.0.12
Flask==1.1.2
gcloud==0.17.0
google-auth==1.22.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
jws==0.1.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauth2client==3.0.0
Pillow==8.0.0
pipenv==2020.8.13
protobuf==3.13.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pylint==2.6.0
PyMySQL==0.10.1
Pyrebase==3.0.27
python-firebase==1.2
python-jwt==2.0.1
pytz==2020.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.6
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
urllib3==1.25.10
virtualenv==20.0.35
virtualenv-clone==0.5.4
Werkzeug==1.0.1
wrapt==1.12.1
